src/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

cx_Freeze==7.2.10
numpy==2.2.3
pillow==11.1.0
PySide6==6.8.2.1
pywin32==309
